coded new features among which is debt management and system to bypass the online checkout payment gateway when ordering

+++ src/app/api/menu-items/route.js
@@ -1,26 +1,27 @@
-// import {isAdmin} from "@/app/api/auth/[...nextauth]/route";
-import { MenuItem } from "@/app/models/MenuItem";
+
+import {isAdmin} from "../auth/[...nextauth]/route";
+
 import mongoose from "mongoose";
+import {MenuItem} from "../../models/MenuItem";
 
 export async function POST(req) {
   mongoose.connect(process.env.MONGO_URL);
   const data = await req.json();
-  // if (await isAdmin()) {
+  if (await isAdmin()) {
     const menuItemDoc = await MenuItem.create(data);
     return Response.json(menuItemDoc);
-  // } else {
-    // return Response.json({});
-  // }
+  } else {
+    return Response.json({});
+  }
 }
 
 export async function PUT(req) {
   mongoose.connect(process.env.MONGO_URL);
-  console.log('req', req)
-  // if (await isAdmin()) {
+  if (await isAdmin()) {
     const {_id, ...data} = await req.json();
     console.log('data', data)
     await MenuItem.findByIdAndUpdate(_id, data, {upsert:true});
-  // }
+  }
   return Response.json(true);
 }
 
@@ -35,8 +36,8 @@ export async function DELETE(req) {
   mongoose.connect(process.env.MONGO_URL);
   const url = new URL(req.url);
   const _id = url.searchParams.get('_id');
-  // if (await isAdmin()) {
+  if (await isAdmin()) {
     await MenuItem.deleteOne({_id});
-  // }
+  }
   return Response.json(true);
 }
